refactor: 테스트를 위한 Mock과 Util 클래스 패키지 분리
