+++ src/test/java/racingcar/mock/MockFixedNumberGenerator.java
@@ -1,4 +1,6 @@
-package racingcar.domain;
+package racingcar.mock;
+
+import racingcar.domain.game.NumberGenerator;
 
 public class MockFixedNumberGenerator implements NumberGenerator {
