+++ src/test/java/racingcar/mock/MoveMethodCalledCountStoreCars.java
@@ -1,0 +1,25 @@
+package racingcar.mock;
+
+import racingcar.domain.car.Cars;
+import racingcar.domain.game.NumberGenerator;
+
+import java.util.List;
+
+public class MoveMethodCalledCountStoreCars extends Cars {
+
+    private int count = 0;
+
+    public MoveMethodCalledCountStoreCars(List<String> cars) {
+        super(cars);
+    }
+
+    @Override
+    public void move(NumberGenerator generator) {
+        super.move(generator);
+        count++;
+    }
+
+    public int getCallCount() {
+        return count;
+    }
+}
